refactor: move controller tests under controller folder
