+++ plasma_framework/test/src/exits/payment/controllers/PaymentPiggybackInFlightExitOnOutput.test.js
@@ -21,13 +21,13 @@ const {
 } = require('openzeppelin-test-helpers');
 const { expect } = require('chai');
 
-const { calculateNormalExitable } = require('../../../helpers/exitable.js');
-const { buildUtxoPos, utxoPosToTxPos } = require('../../../helpers/positions.js');
-const { buildOutputGuard } = require('../../../helpers/utils.js');
-const { PaymentTransactionOutput, PaymentTransaction } = require('../../../helpers/transaction.js');
+const { calculateNormalExitable } = require('../../../../helpers/exitable.js');
+const { buildUtxoPos, utxoPosToTxPos } = require('../../../../helpers/positions.js');
+const { buildOutputGuard } = require('../../../../helpers/utils.js');
+const { PaymentTransactionOutput, PaymentTransaction } = require('../../../../helpers/transaction.js');
 const {
     PROTOCOL, TX_TYPE, VAULT_ID, SAFE_GAS_STIPEND, EMPTY_BYTES_32,
-} = require('../../../helpers/constants.js');
+} = require('../../../../helpers/constants.js');
 
 contract('PaymentPiggybackInFlightExitOnOutput', ([_, alice, inputOwner, outputOwner, nonOutputOwner]) => {
     const ETH = constants.ZERO_ADDRESS;
