+++ plasma_framework/test/src/exits/payment/controllers/PaymentStartInFlightExit.test.js
@@ -24,12 +24,12 @@ const {
 } = require('openzeppelin-test-helpers');
 const { expect } = require('chai');
 
-const { buildUtxoPos, UtxoPos } = require('../../../helpers/positions.js');
-const { computeNormalOutputId, spentOnGas } = require('../../../helpers/utils.js');
-const { PROTOCOL, VAULT_ID, SAFE_GAS_STIPEND } = require('../../../helpers/constants.js');
+const { buildUtxoPos, UtxoPos } = require('../../../../helpers/positions.js');
+const { computeNormalOutputId, spentOnGas } = require('../../../../helpers/utils.js');
+const { PROTOCOL, VAULT_ID, SAFE_GAS_STIPEND } = require('../../../../helpers/constants.js');
 const {
     buildValidIfeStartArgs, buildIfeStartArgs, createInputTransaction, createDepositTransaction, createInFlightTx,
-} = require('../../../helpers/ife.js');
+} = require('../../../../helpers/ife.js');
 
 contract('PaymentStartInFlightExit', ([_, alice, richFather, carol]) => {
     const CHILD_BLOCK_INTERVAL = 1000;
